refactor: toLocalDateString を lib/date-utils に移動 (#857)
UIコンポーネント (session-calendar.tsx) に同居していた日付変換ユーティリティを
専用モジュール (lib/date-utils.ts) に移動し、インポートパスを更新。
