+++ components/calendar/session-calendar-hover.test.ts
@@ -1,9 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  buildSessionDates,
-  getDayCellClassNames,
-  toLocalDateString,
-} from "./session-calendar";
+import { buildSessionDates, getDayCellClassNames } from "./session-calendar";
 
 describe("buildSessionDates", () => {
   it("events が undefined → 空 Set を返す", () => {
@@ -120,15 +116,3 @@ describe("getDayCellClassNames", () => {
   });
 });
 
-describe("toLocalDateString", () => {
-  it("Date オブジェクトからローカル日付の YYYY-MM-DD 文字列を返す", () => {
-    // 日中の時刻: タイムゾーンに依存しない
-    const date = new Date(2025, 0, 15, 14, 0);
-    expect(toLocalDateString(date)).toBe("2025-01-15");
-  });
-
-  it("月・日が1桁の場合もゼロ埋めされる", () => {
-    const date = new Date(2025, 2, 5, 10, 0);
-    expect(toLocalDateString(date)).toBe("2025-03-05");
-  });
-});
