+++ lib/date-utils.ts
@@ -32,6 +32,13 @@ export const formatDateTimeForInput = (date: Date) => {
   return `${formatDateForInput(date)}T${timePart}`;
 };
 
+export function toLocalDateString(date: Date): string {
+  const y = date.getFullYear();
+  const m = String(date.getMonth() + 1).padStart(2, "0");
+  const d = String(date.getDate()).padStart(2, "0");
+  return `${y}-${m}-${d}`;
+}
+
 export function formatTooltipDateTime(
   startsAt: string | Date,
   endsAt: string | Date,
